+++ src/lib.rs
@@ -2,6 +2,10 @@
 //!
 //! Functions for creating time-lagged versions of time series data.
 //!
+//! ## Crate Features
+//!
+//! * `ndarray` - Enables support for [ndarray](https://crates.io/crates/ndarray)'s `Array1` and `Array2` traits.
+//!
 //! ## Example
 //!
 //! For singular time series:
@@ -42,7 +46,7 @@
 //! let lag = f64::INFINITY;
 //! let padding = f64::INFINITY;
 //!
-//! let lagged = lag_matrix_2d(&data, MatrixLayout::RowWise(4), 3, lag, 5).unwrap();
+//! let lagged = lag_matrix_2d(&data, MatrixLayout::RowMajor(4), 3, lag, 5).unwrap();
 //!
 //! assert_eq!(
 //!     lagged,
@@ -75,7 +79,7 @@
 //! let padding = f64::INFINITY;
 //!
 //! // Example row stride of nine: 2 time series × (1 original + 3 lags) + 1 extra padding.
-//! let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnWise(4), 3, lag, 9).unwrap();
+//! let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnMajor(4), 3, lag, 9).unwrap();
 //!
 //! assert_eq!(
 //!     lagged,
@@ -101,14 +105,28 @@
 // the `docsrs` configuration attribute is defined
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
+#[cfg(feature = "ndarray")]
+#[cfg_attr(docsrs, doc(cfg(feature = "ndarray")))]
+mod ndarray_support;
+
 use std::borrow::Borrow;
 use std::fmt::{Display, Formatter};
+
+#[cfg(feature = "ndarray")]
+#[cfg_attr(docsrs, doc(cfg(feature = "ndarray")))]
+pub use ndarray_support::LagMatrixFromArray;
 
 /// The prelude.
 pub mod prelude {
     pub use crate::CreateLagMatrix;
+
+    #[cfg(feature = "ndarray")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "ndarray")))]
+    pub use crate::ndarray_support::LagMatrixFromArray;
 }
 
+/// Provides the [`lag_matrix`](CreateLagMatrix::lag_matrix) and [`lag_matrix_2d`](CreateLagMatrix::lag_matrix_2d)
+/// functions for slice-able copy-able types.
 pub trait CreateLagMatrix<T> {
     /// Create a time-lagged matrix of time series values.
     ///
@@ -201,7 +219,7 @@ pub trait CreateLagMatrix<T> {
     /// let lag = f64::INFINITY;
     /// let padding = f64::INFINITY;
     ///
-    /// let lagged = lag_matrix_2d(&data, MatrixLayout::RowWise(4), 3, lag, 5).unwrap();
+    /// let lagged = lag_matrix_2d(&data, MatrixLayout::RowMajor(4), 3, lag, 5).unwrap();
     ///
     /// assert_eq!(
     ///     lagged,
@@ -234,7 +252,7 @@ pub trait CreateLagMatrix<T> {
     /// let padding = f64::INFINITY;
     ///
     /// // Example row stride of nine: 2 time series × (1 original + 3 lags) + 1 extra padding.
-    /// let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnWise(4), 3, lag, 9).unwrap();
+    /// let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnMajor(4), 3, lag, 9).unwrap();
     ///
     /// assert_eq!(
     ///     lagged,
@@ -370,19 +388,19 @@ pub enum MatrixLayout {
     /// and the columns represent points in time.
     ///
     /// The values represents the number of elements per row, i.e. the length of each time series.
-    RowWise(usize),
+    RowMajor(usize),
     /// Data is laid out column-wise, i.e. reach column of the matrix contains a time series
     /// and the rows represent points in time.
     ///
     /// The values represents the number of elements per column, i.e. the length of each time series.
-    ColumnWise(usize),
+    ColumnMajor(usize),
 }
 
 impl MatrixLayout {
     pub fn len(&self) -> usize {
         match self {
-            MatrixLayout::RowWise(len) => *len,
-            MatrixLayout::ColumnWise(len) => *len,
+            MatrixLayout::RowMajor(len) => *len,
+            MatrixLayout::ColumnMajor(len) => *len,
         }
     }
 }
@@ -432,7 +450,7 @@ impl MatrixLayout {
 /// let lag = f64::INFINITY;
 /// let padding = f64::INFINITY;
 ///
-/// let lagged = lag_matrix_2d(&data, MatrixLayout::RowWise(4), 3, lag, 5).unwrap();
+/// let lagged = lag_matrix_2d(&data, MatrixLayout::RowMajor(4), 3, lag, 5).unwrap();
 ///
 /// assert_eq!(
 ///     lagged,
@@ -465,7 +483,7 @@ impl MatrixLayout {
 /// let padding = f64::INFINITY;
 ///
 /// // Example row stride of nine: 2 time series × (1 original + 3 lags) + 1 extra padding.
-/// let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnWise(4), 3, lag, 9).unwrap();
+/// let lagged = lag_matrix_2d(&data, MatrixLayout::ColumnMajor(4), 3, lag, 9).unwrap();
 ///
 /// assert_eq!(
 ///     lagged,
@@ -509,7 +527,7 @@ pub fn lag_matrix_2d<T: Copy>(
     }
 
     Ok(match layout {
-        MatrixLayout::RowWise(_) => {
+        MatrixLayout::RowMajor(_) => {
             if row_stride < series_length {
                 return Err(LagError::InvalidStride);
             }
@@ -530,7 +548,7 @@ pub fn lag_matrix_2d<T: Copy>(
             }
             lagged
         }
-        MatrixLayout::ColumnWise(_) => {
+        MatrixLayout::ColumnMajor(_) => {
             if row_stride < num_series * lags {
                 return Err(LagError::InvalidStride);
             }
@@ -567,6 +585,7 @@ pub fn lag_matrix_2d<T: Copy>(
     })
 }
 
+/// An error during creation of a lagged data matrix.
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub enum LagError {
     /// The number of lags is greater than the number of data points.
@@ -575,6 +594,8 @@ pub enum LagError {
     InvalidStride,
     /// The number of data points does not match the row/column length specified.
     InvalidLength,
+    /// The data is in an invalid (e.g. non-contiguous) memory layout.
+    InvalidMemoryLayout,
 }
 
 impl std::error::Error for LagError {}
@@ -597,6 +618,10 @@ impl Display for LagError {
             LagError::InvalidLength => write!(
                 f,
                 "The number of data points does not match the row/column length specified"
+            ),
+            LagError::InvalidMemoryLayout => write!(
+                f,
+                "The data is in an invalid (e.g. non-contiguous) memory layout"
             ),
         }
     }
@@ -679,7 +704,7 @@ mod tests {
         let lag = f64::INFINITY;
         let padding = f64::INFINITY;
 
-        let direct = lag_matrix_2d(&data, MatrixLayout::RowWise(4), 3, lag, 5).unwrap();
+        let direct = lag_matrix_2d(&data, MatrixLayout::RowMajor(4), 3, lag, 5).unwrap();
 
         assert_eq!(
             direct,
@@ -710,7 +735,7 @@ mod tests {
         let lag = f64::INFINITY;
         let padding = f64::INFINITY;
 
-        let direct = lag_matrix_2d(&data, MatrixLayout::ColumnWise(4), 3, lag, 9).unwrap();
+        let direct = lag_matrix_2d(&data, MatrixLayout::ColumnMajor(4), 3, lag, 9).unwrap();
 
         assert_eq!(
             direct,